refactor: move cfas routes to organisme
