+++ server/src/http/server.js
@@ -19,7 +19,6 @@ import indicateursExportRouter from "./routes/specific.routes/indicateurs-export
 import dossierApprenantRouter from "./routes/specific.routes/dossiers-apprenants.routes.js";
 import loginRouter from "./routes/specific.routes/old/login.route.js";
 import referentielRouter from "./routes/specific.routes/old/referentiel.route.js";
-import cfasRouter from "./routes/specific.routes/old/cfas.route.js";
 import organismesRouter from "./routes/specific.routes/organismes.routes.js";
 import formationRouter from "./routes/specific.routes/old/formations.route.js";
 import indicateursNationalRouter from "./routes/specific.routes/indicateurs-national.routes.js";
@@ -156,7 +155,6 @@ export default async (services) => {
   ); // EXPOSED TO REFERENTIEL PROTECTED BY API KEY
 
   app.use("/api/formations", formationRouter()); // FRONT
-  app.use("/api/cfas", cfasRouter()); // FRONT
   app.use("/api/referentiel", referentielRouter()); // FRONT
 
   // The error handler must be before any other error middleware and after all controllers
